+++ src/usr/host.rs
@@ -139,7 +139,7 @@ pub fn resolve(name: &str) -> Result<IpAddress, ResponseCode> {
         return Err(ResponseCode::NetworkError);
     };
 
-    let flags = OpenFlag::Device as usize;
+    let flags = OpenFlag::Device as u8;
     if let Some(handle) = syscall::open(socket_path, flags) {
         if syscall::connect(handle, addr, port).is_err() {
             syscall::close(handle);
